+++ blotztask-mobile/src/app/(protected)/(tabs)/settings/review.tsx
@@ -1,4 +1,4 @@
-import MonthlyReviewScreen from "@/feature/monthly-review/screens/monthly-review-screen";
+import MonthlyReviewScreen from "@/feature/review/screens/monthly-review-screen";
 
 export default function ReviewRoute() {
   return <MonthlyReviewScreen />;
